Change the cranelift-wasm FuncEnvironment interface to make it possible for the TargetEnvironment to have the values on the wasm stack not match the values declared in wasm, on function entry and exit (and similarly around calls).
translate_call and translate_call_indirect are changed to return a vector of values they've unpacked the returned values into, instead of a cranelift IR instruction.

When initializing the wasm value stack, pass in the number of values expected
as output from the exit block, since it may not match the number of values
declared in the cranelift function signature.